refactor(inventory): split core module into package and move display to UI
Extract inventory logic into submodules and keep presentation in ui/menus/_display.
