--- core/inventory.py
@@ -1,658 +1,0 @@
-"""Инвентарь персонажа, экипировка и расчёт КД."""
-
-import re
-from typing import Any, Literal
-
-from core.classes import character_has_spellcasting
-from core.dice import ability_modifier
-from core.equipment import (
-    armor_category,
-    default_ammunition_pack_size,
-    get_armor_name,
-    get_equipment_item_name,
-    get_tool_name,
-    get_weapon_name,
-    load_armor,
-    load_equipment_item,
-    load_weapon,
-    meets_armor_strength_requirement,
-    weapon_ammunition_item_id,
-    weapon_category,
-    weapon_range,
-)
-from core.feat_apply import (
-    dual_wielder_ac_bonus_from_feats,
-    has_non_light_dual_wield,
-)
-from core.models import Character
-from core.proficiency_checks import (
-    has_armor_proficiency,
-    has_weapon_proficiency,
-)
-from core.stats import ABILITY_SCORE_DEFAULT
-from core.types import StringsDict
-
-ItemKind = Literal["weapon", "armor", "tool", "equipment"]
-
-
-def normalize_inventory_item(raw: dict[str, Any]) -> dict[str, Any] | None:
-    """Нормализовать запись инвентаря."""
-    kind = raw.get("kind")
-    item_id = raw.get("id")
-    if kind not in ("weapon", "armor", "tool", "equipment"):
-        return None
-    if not isinstance(item_id, str) or not item_id:
-        return None
-    qty_raw = raw.get("qty", 1)
-    qty = int(qty_raw) if isinstance(qty_raw, int) else 1
-    if qty < 1:
-        qty = 1
-    return {"kind": kind, "id": item_id, "qty": qty}
-
-
-def expand_pack_contents(item_id: str) -> list[dict[str, Any]]:
-    """Развернуть набор (pack) в список предметов."""
-    info = load_equipment_item(item_id)
-    if info.get("category") != "pack":
-        return []
-    contents = info.get("contents", [])
-    if not isinstance(contents, list):
-        return []
-    result: list[dict[str, Any]] = []
-    for entry in contents:
-        if not isinstance(entry, dict):
-            continue
-        normalized = normalize_inventory_item(entry)
-        if normalized:
-            result.append(normalized)
-    return result
-
-
-def merge_inventory_items(
-    items: list[dict[str, Any]],
-) -> list[dict[str, Any]]:
-    """Сложить одинаковые предметы по kind+id."""
-    merged: dict[tuple[str, str], int] = {}
-    order: list[tuple[str, str]] = []
-    for raw in items:
-        item = normalize_inventory_item(raw) if "kind" in raw else None
-        if item is None:
-            continue
-        key = (str(item["kind"]), str(item["id"]))
-        if key not in merged:
-            order.append(key)
-            merged[key] = 0
-        merged[key] += int(item["qty"])
-    return [
-        {"kind": kind, "id": item_id, "qty": merged[(kind, item_id)]}
-        for kind, item_id in order
-    ]
-
-
-def add_items_to_inventory(
-    inventory: list[dict[str, Any]],
-    new_items: list[dict[str, Any]],
-) -> list[dict[str, Any]]:
-    """Добавить предметы в инвентарь с развёрткой наборов."""
-    expanded: list[dict[str, Any]] = list(inventory)
-    for raw in new_items:
-        item = normalize_inventory_item(raw)
-        if item is None:
-            continue
-        if (
-            item["kind"] == "equipment"
-            and load_equipment_item(item["id"]).get("category") == "pack"
-        ):
-            expanded.extend(expand_pack_contents(item["id"]))
-        else:
-            expanded.append(item)
-    return merge_inventory_items(expanded)
-
-
-def default_equipped() -> dict[str, Any]:
-    """Пустая экипировка."""
-    return {
-        "armor": None,
-        "shield": False,
-        "main_hand": None,
-        "off_hand": None,
-    }
-
-
-def _armor_ac_value(armor_id: str, dex_mod: int) -> int:
-    """Базовый КД от доспеха по PHB."""
-    info = load_armor(armor_id)
-    base = int(info.get("armor_class", 10))
-    bonus_type = str(info.get("modifier_bonus", ""))
-    max_dex = info.get("max_dex_modifier")
-    dex_part = 0
-    if bonus_type == "DEX":
-        if isinstance(max_dex, int):
-            dex_part = min(dex_mod, max_dex)
-        else:
-            dex_part = dex_mod
-    return base + dex_part
-
-
-def compute_ac(character: Character) -> int:
-    """Класс доспеха персонажа (PHB + бонусы черт при экипировке)."""
-    stats = character.stats
-    dexterity = int(stats.get("dexterity", ABILITY_SCORE_DEFAULT))
-    dex_mod = ability_modifier(dexterity)
-    equipped = character.equipped or default_equipped()
-    armor_id = equipped.get("armor")
-    ac = 10 + dex_mod
-    if isinstance(armor_id, str) and armor_id:
-        ac = _armor_ac_value(armor_id, dex_mod)
-    if equipped.get("shield"):
-        ac += 2
-    if dual_wielder_ac_bonus_applies(equipped, character.feat_ids):
-        ac += dual_wielder_ac_bonus_from_feats(character.feat_ids)
-    return ac
-
-
-def _armor_sort_key(armor_id: str) -> tuple[int, int]:
-    """Ключ сортировки: тяжелее и выше базовый КД — лучше."""
-    info = load_armor(armor_id)
-    cat = armor_category(armor_id)
-    cat_rank = {"heavy": 3, "medium": 2, "light": 1}.get(cat, 0)
-    return (cat_rank, int(info.get("armor_class", 0)))
-
-
-def _parse_dice_average(dice: str) -> float:
-    """Средний урон по нотации кости (например ``1d8`` → 4.5)."""
-    match = re.fullmatch(r"(\d+)d(\d+)", dice.strip())
-    if not match:
-        return 0.0
-    count, sides = int(match.group(1)), int(match.group(2))
-    return count * (sides + 1) / 2
-
-
-def _weapon_properties(weapon_id: str) -> dict[str, Any]:
-    props = load_weapon(weapon_id).get("properties")
-    return dict(props) if isinstance(props, dict) else {}
-
-
-def _weapon_is_two_handed(weapon_id: str) -> bool:
-    return bool(_weapon_properties(weapon_id).get("two_handed"))
-
-
-def _weapon_is_light(weapon_id: str) -> bool:
-    return bool(_weapon_properties(weapon_id).get("light"))
-
-
-def _weapon_is_melee(weapon_id: str) -> bool:
-    return weapon_category(weapon_id).endswith("_melee")
-
-
-def _weapon_is_one_handed_melee(weapon_id: str) -> bool:
-    return _weapon_is_melee(weapon_id) and not _weapon_is_two_handed(weapon_id)
-
-
-def dual_wielder_ac_bonus_applies(
-    equipped: dict[str, Any], feat_ids: list[str]
-) -> bool:
-    """+1 КД: в каждой руке одноручное рукопашное (черта dual_wielder)."""
-    if not feat_ids or dual_wielder_ac_bonus_from_feats(feat_ids) <= 0:
-        return False
-    if equipped.get("shield"):
-        return False
-    main = equipped.get("main_hand")
-    off = equipped.get("off_hand")
-    if not isinstance(main, str) or not main:
-        return False
-    if not isinstance(off, str) or not off:
-        return False
-    if main_hand_uses_both_hands(equipped):
-        return False
-    return _weapon_is_one_handed_melee(main) and _weapon_is_one_handed_melee(
-        off
-    )
-
-
-def _weapon_base_damage_dice(weapon_id: str) -> str:
-    damage = load_weapon(weapon_id).get("damage", {})
-    if isinstance(damage, dict):
-        return str(damage.get("dice", "1d4"))
-    return "1d4"
-
-
-def _weapon_is_versatile(weapon_id: str) -> bool:
-    return bool(_weapon_properties(weapon_id).get("versatile"))
-
-
-def _weapon_versatile_damage_dice(weapon_id: str) -> str:
-    versatile = _weapon_properties(weapon_id).get("versatile")
-    if versatile:
-        return str(versatile)
-    return _weapon_base_damage_dice(weapon_id)
-
-
-def _weapon_one_handed_damage(weapon_id: str) -> float:
-    """Средний урон одной рукой (без versatile в двухручном режиме)."""
-    if _weapon_is_two_handed(weapon_id):
-        return 0.0
-    return _parse_dice_average(_weapon_base_damage_dice(weapon_id))
-
-
-def _weapon_auto_equip_damage(weapon_id: str) -> tuple[float, bool]:
-    """Урон для авто-экипировки и флаг «нативно двуручное»."""
-    if _weapon_is_two_handed(weapon_id):
-        return _parse_dice_average(_weapon_base_damage_dice(weapon_id)), True
-    if _weapon_is_versatile(weapon_id):
-        dice = _weapon_versatile_damage_dice(weapon_id)
-        return _parse_dice_average(dice), False
-    return _weapon_one_handed_damage(weapon_id), False
-
-
-def main_hand_uses_both_hands(equipped: dict[str, Any]) -> bool:
-    """Основное оружие занимает обе руки (двуручное или универсальное)."""
-    main = equipped.get("main_hand")
-    if not isinstance(main, str) or not main:
-        return False
-    if _weapon_is_two_handed(main):
-        return True
-    if not _weapon_is_versatile(main):
-        return False
-    grip = equipped.get("main_hand_grip")
-    if grip == "two_handed":
-        return True
-    if grip == "one_handed":
-        return False
-    return not equipped.get("shield") and not equipped.get("off_hand")
-
-
-def _pick_main_weapon(weapons: list[str]) -> tuple[str | None, bool]:
-    """Основное оружие с наибольшим уроном (с учётом владения)."""
-    if not weapons:
-        return None, False
-    best_id = weapons[0]
-    best_damage = -1.0
-    best_two_handed = False
-    for weapon_id in weapons:
-        damage, two_handed = _weapon_auto_equip_damage(weapon_id)
-        if damage > best_damage:
-            best_damage = damage
-            best_id = weapon_id
-            best_two_handed = two_handed
-    return best_id, best_two_handed
-
-
-def _inventory_weapon_ids(character: Character) -> list[str]:
-    """Уникальные id оружия в инвентаре."""
-    seen: set[str] = set()
-    result: list[str] = []
-    for item in character.inventory:
-        if item.get("kind") != "weapon":
-            continue
-        weapon_id = str(item.get("id", ""))
-        if not weapon_id or weapon_id in seen:
-            continue
-        seen.add(weapon_id)
-        result.append(weapon_id)
-    return result
-
-
-def _proficient_inventory_weapons(character: Character) -> list[str]:
-    """Оружие из инвентаря, которым владеет персонаж."""
-    profs = character.weapon_proficiencies
-    return [
-        weapon_id
-        for weapon_id in _inventory_weapon_ids(character)
-        if has_weapon_proficiency(profs, weapon_id)
-    ]
-
-
-def _off_hand_weapon_allowed(weapon_id: str, *, allow_non_light: bool) -> bool:
-    """Одноручное рукопашное для второй руки (лёгкое или dual_wielder)."""
-    if not _weapon_is_one_handed_melee(weapon_id):
-        return False
-    if allow_non_light:
-        return True
-    return _weapon_is_light(weapon_id)
-
-
-def _pick_off_hand_weapon(
-    weapons: list[str],
-    main_hand: str,
-    *,
-    allow_non_light: bool = False,
-) -> str | None:
-    """Одноручное оружие для второй руки (кроме основного)."""
-    candidates: list[tuple[float, str]] = []
-    for weapon_id in weapons:
-        if weapon_id == main_hand:
-            continue
-        if not _off_hand_weapon_allowed(
-            weapon_id, allow_non_light=allow_non_light
-        ):
-            continue
-        candidates.append((_weapon_one_handed_damage(weapon_id), weapon_id))
-    if not candidates:
-        return None
-    return max(candidates)[1]
-
-
-def equip_defaults(character: Character) -> dict[str, Any]:
-    """Авто-экипировка: лучший доспех, щит и оружие из инвентаря."""
-    equipped = default_equipped()
-    armor_profs = character.armor_proficiencies
-    strength = int(character.stats.get("strength", ABILITY_SCORE_DEFAULT))
-
-    armor_ids = [
-        str(item["id"])
-        for item in character.inventory
-        if item.get("kind") == "armor"
-        and armor_category(str(item["id"])) in ("light", "medium", "heavy")
-        and has_armor_proficiency(armor_profs, str(item["id"]))
-        and meets_armor_strength_requirement(str(item["id"]), strength)
-    ]
-    if armor_ids:
-        equipped["armor"] = max(armor_ids, key=_armor_sort_key)
-
-    has_shield_item = any(
-        item.get("kind") == "armor" and item.get("id") == "shield"
-        for item in character.inventory
-    )
-    shield_proficient = has_armor_proficiency(armor_profs, "shield")
-    can_use_shield = has_shield_item and shield_proficient
-
-    weapons = _proficient_inventory_weapons(character)
-    main_hand, main_two_handed = _pick_main_weapon(weapons)
-    if main_hand:
-        equipped["main_hand"] = main_hand
-
-    if main_hand and main_two_handed:
-        equipped["shield"] = False
-        return equipped
-
-    if can_use_shield:
-        equipped["shield"] = True
-        if main_hand and _weapon_is_versatile(main_hand):
-            equipped["main_hand_grip"] = "one_handed"
-        return equipped
-
-    if character_has_spellcasting(
-        character.class_id, character.subclass_id, character.level
-    ):
-        return equipped
-
-    allow_non_light = has_non_light_dual_wield(character.feat_ids)
-    off_hand = _pick_off_hand_weapon(
-        weapons,
-        main_hand or "",
-        allow_non_light=allow_non_light,
-    )
-    if off_hand:
-        equipped["off_hand"] = off_hand
-        if main_hand and _weapon_is_versatile(main_hand):
-            equipped["main_hand_grip"] = "one_handed"
-        return equipped
-
-    if main_hand and _weapon_is_versatile(main_hand):
-        equipped["main_hand_grip"] = "two_handed"
-    return equipped
-
-
-def item_display_name(
-    kind: str,
-    item_id: str,
-    language: str = "ru",
-) -> str:
-    """Локализованное имя предмета инвентаря."""
-    if kind == "weapon":
-        return get_weapon_name(item_id, language)
-    if kind == "armor":
-        return get_armor_name(item_id, language)
-    if kind == "tool":
-        return get_tool_name(item_id, language)
-    return get_equipment_item_name(item_id, language)
-
-
-def _equipped_item_counts(
-    equipped: dict[str, Any],
-) -> dict[tuple[str, str], int]:
-    """Сколько предметов каждого kind+id занято экипировкой."""
-    counts: dict[tuple[str, str], int] = {}
-    armor_id = equipped.get("armor")
-    if isinstance(armor_id, str) and armor_id:
-        key = ("armor", armor_id)
-        counts[key] = counts.get(key, 0) + 1
-    if equipped.get("shield"):
-        key = ("armor", "shield")
-        counts[key] = counts.get(key, 0) + 1
-    main_hand = equipped.get("main_hand")
-    if isinstance(main_hand, str) and main_hand:
-        key = ("weapon", main_hand)
-        counts[key] = counts.get(key, 0) + 1
-    off_hand = equipped.get("off_hand")
-    if isinstance(off_hand, str) and off_hand:
-        key = ("weapon", off_hand)
-        counts[key] = counts.get(key, 0) + 1
-    return counts
-
-
-def inventory_excluding_equipped(
-    inventory: list[dict[str, Any]],
-    equipped: dict[str, Any] | None,
-) -> list[dict[str, Any]]:
-    """Инвентарь для UI: без экипированного; save не меняется."""
-    if not equipped:
-        return list(inventory)
-    hidden = _equipped_item_counts(equipped)
-    if not hidden:
-        return list(inventory)
-    visible: list[dict[str, Any]] = []
-    for item in inventory:
-        kind = str(item.get("kind", ""))
-        item_id = str(item.get("id", ""))
-        qty = int(item.get("qty", 1))
-        remaining = qty - hidden.get((kind, item_id), 0)
-        if remaining > 0:
-            visible.append({"kind": kind, "id": item_id, "qty": remaining})
-    return visible
-
-
-def format_inventory_line(
-    inventory: list[dict[str, Any]],
-    language: str = "ru",
-    *,
-    equipped: dict[str, Any] | None = None,
-) -> str:
-    """Сжатый список инвентаря для UI (без экипированных предметов)."""
-    display_items = inventory_excluding_equipped(inventory, equipped)
-    parts: list[str] = []
-    for item in display_items:
-        kind = str(item.get("kind", ""))
-        item_id = str(item.get("id", ""))
-        qty = int(item.get("qty", 1))
-        name = item_display_name(kind, item_id, language)
-        if qty > 1:
-            parts.append(f"{name} ×{qty}")
-        else:
-            parts.append(name)
-    return ", ".join(parts)
-
-
-def _versatile_can_switch_to_two_hands(equipped: dict[str, Any]) -> bool:
-    """Можно ли взять универсальное оружие двумя руками."""
-    if equipped.get("shield"):
-        return False
-    off = equipped.get("off_hand")
-    return not (isinstance(off, str) and off)
-
-
-def _versatile_active_grip(
-    equipped: dict[str, Any], weapon_id: str
-) -> Literal["one_handed", "two_handed"]:
-    """Текущий режим универсального оружия."""
-    grip = equipped.get("main_hand_grip")
-    if grip == "one_handed":
-        return "one_handed"
-    if grip == "two_handed":
-        return "two_handed"
-    if main_hand_uses_both_hands(equipped):
-        return "two_handed"
-    return "one_handed"
-
-
-def _versatile_grip_hint(
-    weapon_id: str,
-    equipped: dict[str, Any],
-    strings: StringsDict,
-) -> str:
-    """Подсказка хвата универсального оружия (одна / две руки)."""
-    from core.localization import get_string
-
-    if not _weapon_is_versatile(weapon_id):
-        return ""
-    if _versatile_active_grip(equipped, weapon_id) == "two_handed":
-        key = "choose_character.field_equipped_versatile_grip_two"
-    else:
-        key = "choose_character.field_equipped_versatile_grip_one"
-    return get_string(strings, key)
-
-
-def format_versatile_damage_dice(
-    weapon_id: str,
-    equipped: dict[str, Any],
-    strings: StringsDict,
-    language: str = "ru",
-) -> tuple[str, str, str] | None:
-    """Кости универсального оружия и активный режим: (1к8, 1к10, one|two)."""
-    from core.equipment import (
-        _format_dice_for_display,
-        _weapon_damage_dice,
-        _weapon_versatile_dice,
-    )
-
-    if not _weapon_is_versatile(weapon_id):
-        return None
-    one_dice = _format_dice_for_display(
-        _weapon_damage_dice(weapon_id), language
-    )
-    two_dice = _format_dice_for_display(
-        _weapon_versatile_dice(weapon_id), language
-    )
-    active = (
-        "two"
-        if _versatile_active_grip(equipped, weapon_id) == "two_handed"
-        else "one"
-    )
-    return one_dice, two_dice, active
-
-
-def inventory_item_quantity(
-    inventory: list[dict[str, Any]], kind: str, item_id: str
-) -> int:
-    """Суммарное количество предмета kind+id в инвентаре."""
-    total = 0
-    for item in inventory:
-        if item.get("kind") == kind and item.get("id") == item_id:
-            total += int(item.get("qty", 1))
-    return total
-
-
-def _loaded_ammunition_qty(
-    inventory: list[dict[str, Any]], ammo_item_id: str
-) -> int:
-    """Боеприпасы «под рукой»: из инвентаря, до ёмкости колчана/сумки."""
-    in_inv = inventory_item_quantity(inventory, "equipment", ammo_item_id)
-    pack = default_ammunition_pack_size(ammo_item_id)
-    return min(in_inv, pack) if in_inv > 0 else 0
-
-
-def _equipped_weapon_for_range(equipped: dict[str, Any]) -> str | None:
-    """Оружие для строки дистанции (основная рука, затем вторая)."""
-    for slot in ("main_hand", "off_hand"):
-        weapon_id = equipped.get(slot)
-        if isinstance(weapon_id, str) and weapon_range(weapon_id):
-            return weapon_id
-    return None
-
-
-def get_equipped_display(
-    character: Character,
-    language: str = "ru",
-) -> tuple[dict[str, str], bool]:
-    """Локализованные подписи экипировки; bool — серая подпись второй руки."""
-    from core.equipment import armor_equipped_hint, weapon_property_hint
-    from core.localization import get_string, load_strings
-
-    strings = load_strings(language)
-    empty = get_string(strings, "choose_character.field_equipped_empty")
-    equipped = character.equipped or default_equipped()
-    result: dict[str, str] = {}
-    off_hand_muted = False
-    armor_id = equipped.get("armor")
-    if isinstance(armor_id, str) and armor_id:
-        result["armor"] = item_display_name("armor", armor_id, language)
-        armor_hint = armor_equipped_hint(armor_id, strings, language)
-        if armor_hint:
-            result["armor_hint"] = armor_hint
-    else:
-        result["armor"] = empty
-    main_hand = equipped.get("main_hand")
-    if isinstance(main_hand, str) and main_hand:
-        result["main_hand"] = item_display_name("weapon", main_hand, language)
-        hint_parts: list[str] = []
-        prop_hint = weapon_property_hint(main_hand, strings, language)
-        if prop_hint:
-            hint_parts.append(prop_hint)
-        grip_hint = _versatile_grip_hint(main_hand, equipped, strings)
-        if grip_hint:
-            hint_parts.append(grip_hint)
-        if hint_parts:
-            result["main_hand_hint"] = ", ".join(hint_parts)
-        damage_dice = format_versatile_damage_dice(
-            main_hand, equipped, strings, language
-        )
-        if damage_dice:
-            one, two, active = damage_dice
-            result["damage_one_dice"] = one
-            result["damage_two_dice"] = two
-            result["damage_active"] = active
-    else:
-        result["main_hand"] = empty
-    if main_hand_uses_both_hands(equipped) and isinstance(main_hand, str):
-        if _weapon_is_two_handed(main_hand):
-            label_key = "choose_character.field_equipped_off_two_handed"
-        else:
-            label_key = "choose_character.field_equipped_off_versatile"
-        result["off_hand"] = get_string(strings, label_key)
-        off_hand_muted = True
-    elif equipped.get("shield"):
-        result["off_hand"] = get_armor_name("shield", language)
-    else:
-        off_hand = equipped.get("off_hand")
-        if isinstance(off_hand, str) and off_hand:
-            result["off_hand"] = item_display_name(
-                "weapon", off_hand, language
-            )
-            hint = weapon_property_hint(off_hand, strings, language)
-            if hint:
-                result["off_hand_hint"] = hint
-        else:
-            result["off_hand"] = empty
-    ammo_weapon = main_hand if isinstance(main_hand, str) else None
-    if ammo_weapon:
-        ammo_id = weapon_ammunition_item_id(ammo_weapon)
-        if ammo_id:
-            qty = _loaded_ammunition_qty(character.inventory, ammo_id)
-            result["ammunition"] = get_string(
-                strings,
-                "choose_character.field_equipped_ammunition_value",
-                type=get_equipment_item_name(ammo_id, language),
-                qty=qty,
-            )
-    range_weapon = _equipped_weapon_for_range(equipped)
-    if range_weapon:
-        rng = weapon_range(range_weapon)
-        if rng:
-            result["distance"] = get_string(
-                strings,
-                "choose_character.field_equipped_range_value",
-                normal=rng["normal"],
-                long=rng["long"],
-            )
-    return result, off_hand_muted
